requirements.txt
apache-airflow-providers-snowflake==1.3.0
pyarrow==8.0.0
snowflake-sqlalchemy==1.2.4
scikit-learn==1.0.2
jupyter==1.0.0
numpy
matplotlib
seaborn
snowflake-snowpark-python[pandas]==1.0.0
pandas
Sphinx
sphinx-rtd-theme